+++ src/pub/proxy_edit/fields.js
@@ -336,6 +336,10 @@ export const tabs = {
                 label: 'Debug',
                 tooltip: `Luminati request debug info`,
             },
+            debug_logs: {
+                label: 'Debug logs',
+                tooltip: 'Detailed debug logs',
+            },
         },
     },
 };